added comments to potentially make enemies run from player
for the player level, we could do the enemies run away if the players level is greater than 3 * floor level. so enemies on floor 1 would run when the player reaches level 4, level 7 on floor 2, level 10 on floor 3, and 13 on floor 4. I finished the game at level 11 doing the genocide run, so i think this is pretty good honestly